nscale is accepts now also floats (also in scientific notation) - useful for large numbers fixed a bug in wrong units when nscale != 1

+++ pymchelper/run.py
@@ -34,7 +34,7 @@ def main(args=sys.argv[1:]):
     parser.add_argument("output", help='output filename or directory', nargs='?')
     parser.add_argument("--many", help='automatically merge data from various sources', action="store_true")
     parser.add_argument("--nan", help='ignore NaN in averaging', action="store_true")
-    parser.add_argument("-n", '--nscale', help='Scale with number of primaries N.', default=1, type=int)
+    parser.add_argument("-n", '--nscale', help='Scale with number of primaries N.', default=1, type=float)
     parser.add_argument("--converter",
                         help='converters',
                         default=(SHConverters.standard.name,),
